glue-etl/glue-requirements.txt
# Core PySpark and AWS Glue
PySpark==3.4.1
aws-glue-sessions==1.0.0
awswrangler==3.7.0  # For optimized S3/Parquet operations

# Data Processing
pandas==2.0.3  # For UDFs and small-data operations
numpy==1.24.4  # Required by pandas
python-dateutil==2.8.2  # Date parsing

# Logging/Monitoring
boto3==1.28.45  # For AWS service calls
botocore==1.31.45  # boto3 dependency

# Job-Specific Helpers
joblib==1.3.2  # For feature engineering models (if needed)
scikit-learn==1.3.0  # For advanced feature engineering

# Dependency Control
--no-binary :all:  # Force pure-Python wheels